A few conversions from float to int and defining constants as float (rather than double) to reduce warnings with Intel comiler.
git-svn-id: http://svn.code.sf.net/p/petavision/code/trunk@2131 7c0444ca-5550-0410-9008-faecdbe78eb5

+++ src/connections/HyPerConn.hpp
@@ -124,7 +124,7 @@ protected:
    int connId;               // connection id
 
    char * name;
-   float nxp, nyp, nfp;      // size of weight dimensions
+   int nxp, nyp, nfp;      // size of weight dimensions
 
    int numParams;
    PVConnParams * params;
